requirements.txt
aiohttp>=3.9
aiohttp-apispec>=2.2
marshmallow>=3.0,<4.0
ccxt>=4.3
SQLAlchemy>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9
python-dotenv>=1.0
uvloop>=0.19
gunicorn>=21.2
alembic>=1.13
pytest>=8.2
pytest-aiohttp>=1.0
aiosqlite>=0.20
greenlet>=3.2
